Revert "feat(wallets) Move wallets API under customers (#4955)"
This reverts commit ea11bcccd2d9172fa8186348b6c874a15665903b.
