+++ app/controllers/api/v1/customers/wallets_controller.rb
@@ -4,29 +4,7 @@ module Api
   module V1
     module Customers
       class WalletsController < BaseController
-        include WalletActions
-
-        def create
-          wallet_create(customer)
-        end
-
-        def update
-          wallet = customer.wallets.find_by(code: params[:code])
-
-          wallet_update(wallet)
-        end
-
-        def terminate
-          wallet = customer.wallets.find_by(code: params[:code])
-
-          wallet_terminate(wallet)
-        end
-
-        def show
-          wallet = customer.wallets.find_by(code: params[:code])
-
-          wallet_show(wallet)
-        end
+        include WalletIndex
 
         def index
           wallet_index(external_customer_id: customer.external_id)
